Update patch set 12
Patch Set 12:

(1 comment)

I've marked a line that is 82 chars long. And there seem to be a few other egregious outliers (those with url references on them).

IMHO it's fine to have a longer standard line length for doco as opposed to code anyway, but if we want that enforced then it should be documented in this readme (don't think it is currently?).

Patch-set: 12