Update patch set 1
Patch Set 1:

the markdown rules weren't getting applied you these files before (or any new file to the repository - hence the need for change), and a change to the scope of the parser now includes your files.  have a look at https://review.rc.nectar.org.au/#/c/3005/.   it's not a temporary change in that we need all markdown files to be parsed, however the fact that the new parsing rules break with files containing spaces in the name can be fixed with a patch to the parsing rules or (my preference) not using names in spaces.  it's not very unix :)

Patch-set: 1