Update patch set 1
Patch Set 1: Code-Review+2

it looks to me as if this is failing now because the `git diff --name-only HEAD~1` part of the markdown parser is returning the D on the old file, and that's where the error is coming from.  so i've accepted this change anyhow.  that should accept your new file and then we can test on that.

next please do a change to your renamed file and i think that you should not get the same parsing errors.

Patch-set: 1
Label: Code-Review=+2