Update patch set 1
Patch Set 1: Code-Review-1

Hi Chris,
In order for your links to work, use [dashboard][dashboard], the first bit is the hypertext and the second bit links to the URL at the bottom of the page. Otherwise it just shows up as a word in square brackets.
https://daringfireball.net/projects/markdown/syntax#link

In the same branch - modify the links, then use git add -A, git commit --amend

Patch-set: 1
Label: Code-Review=-1