Update patch set 1
Patch Set 1:

(4 comments)

Hi Mark. A few nits here but otherwise ok.

If you're going to make edits it'd be nice to fill out a few of the points with either good references or direct config examples.

Also, I think it'd be worth adding a paragraph at the end mentioning that the "NeCTAR Images" (link to that tier0 doco) come preconfigured with many of these recommendations.

Patch-set: 1
Label: Verified=0